+++ lib/generators_test.ts
@@ -1,7 +1,6 @@
 import {
   assert,
   assertEquals,
-  assertExists,
 } from "https://deno.land/std@0.84.0/testing/asserts.ts";
 import * as generators from "./generators.ts";