+++ lib/transformers.ts
@@ -1,20 +1,29 @@
+import { stripIterable } from "./internal/util.ts";
+
 /**
- * {@link map | `map`} callback
- * @callback MapCallback
- * @param item - The current item to be mapped.
+ * @link map | `map`} callback.
  * @typeParam T - See {@link map}
  * @typeParam U - See {@link map}
  */
 export interface MapCallback<T, U> {
-  (item: T): U;
+  /**
+   * {@link map | `map`} callback.
+   * @callback MapCallback
+   * @param item - The current item to be mapped.
+   * @param index - The index of the item.
+   * @param it - The iterable.
+   * @returns The mapped value
+   */
+  (item: T, index: number, it: Iterable<T>): U;
 }
 
 /**
  * Lazily calls a defined callback function for each element of an iterable, and
  * returns a new iterator of the results.
  * @param it - The iterable being mapped.
- * @param {MapCallback} f - A function which excepts an item of `it` as a single
- * argument. Called for every item of `it`
+ * @param {MapCallback} f - A function that accepts up to three arguments. The
+ * map method calls the callbackfn function one time for each element in the
+ * array.
  * @typeParam T - Type of items in `it`.
  * @typeParam U - Return type of `f`.
  * @returns An iterator of `f` applied to items of `it`.
@@ -23,8 +32,10 @@ export function* map<T, U = T>(
   it: Iterable<T>,
   f: MapCallback<T, U>,
 ): IterableIterator<U> {
-  for (const i of it) {
-    yield f(i);
+  let i = 0;
+  for (const item of it) {
+    yield f(item, i, stripIterable(it));
+    i++;
   }
 }